+++ doc/changelog.rst
@@ -16,8 +16,8 @@ General
 
 MJX
 ^^^
-
 - Added ``mocap_pos`` and ``mocap_quat`` in kinematics.
+- Added support for :ref:`spatial tendons <tendon-spatial>` with external sphere and cylinder wrapping.
 
 Bug fixes
 ^^^^^^^^^